Upload code employee for bakery

+++ src/main/java/model/Payroll.java
@@ -1,6 +1,7 @@
 package model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class Payroll implements Serializable {
 
@@ -8,15 +9,18 @@ public class Payroll implements Serializable {
     private String fullName;
     private int month;
     private int workDays;
-    private double baseSalary;
-    private double bonus;
-    private double penalty;
-    private double totalSalary;
+
+    private BigDecimal baseSalary;
+    private BigDecimal bonus;
+    private BigDecimal penalty;
+    private BigDecimal totalSalary;
 
     public Payroll() {}
 
     public Payroll(String employeeID, String fullName, int month, int workDays,
-                   double baseSalary, double bonus, double penalty, double totalSalary) {
+                   BigDecimal baseSalary, BigDecimal bonus,
+                   BigDecimal penalty, BigDecimal totalSalary) {
+
         this.employeeID = employeeID;
         this.fullName = fullName;
         this.month = month;
@@ -27,12 +31,35 @@ public class Payroll implements Serializable {
         this.totalSalary = totalSalary;
     }
 
-    public String getEmployeeID() { return employeeID; }
-    public String getFullName() { return fullName; }
-    public int getMonth() { return month; }
-    public int getWorkDays() { return workDays; }
-    public double getBaseSalary() { return baseSalary; }
-    public double getBonus() { return bonus; }
-    public double getPenalty() { return penalty; }
-    public double getTotalSalary() { return totalSalary; }
+    public String getEmployeeID() {
+        return employeeID;
+    }
+
+    public String getFullName() {
+        return fullName;
+    }
+
+    public int getMonth() {
+        return month;
+    }
+
+    public int getWorkDays() {
+        return workDays;
+    }
+
+    public BigDecimal getBaseSalary() {
+        return baseSalary;
+    }
+
+    public BigDecimal getBonus() {
+        return bonus;
+    }
+
+    public BigDecimal getPenalty() {
+        return penalty;
+    }
+
+    public BigDecimal getTotalSalary() {
+        return totalSalary;
+    }
 }
